commented code that needs to be refactored
