+++ dork/types.py
@@ -435,35 +435,35 @@ class Game:
         Gamebuilder.save_game(self.hero.name, self.data)
         return "game saved successfully!", False
 
-    def _take(self, item="all"):
-        # Item defaults to "all", and adds all items in room to inventory
-        room_items = self.hero.location.items
-        room_items2 = room_items.copy()
-        player = self.hero.items
-        if item == "all":
-            for item_n in room_items2:
-                player[item_n] = room_items.pop(item_n)
-            return f"You took {item} item. You took them well.", False
-        player[item] = room_items.pop(item)
-        return f"You took the {item}. You took it well.", False
+    # def _take(self, item="all"):
+    #     # Item defaults to "all", and adds all items in room to inventory
+    #     room_items = self.hero.location.items
+    #     room_items2 = room_items.copy()
+    #     player = self.hero.items
+    #     if item == "all":
+    #         for item_n in room_items2:
+    #             player[item_n] = room_items.pop(item_n)
+    #         return f"You took {item} item. You took them well.", False
+    #     player[item] = room_items.pop(item)
+    #     return f"You took the {item}. You took it well.", False
 
-    def _drop_item(self, item):
-        """drops specific item from player to room"""
-        player = self.hero.items
-        room_items = self.hero.location.items
-        room_items[item] = player.pop(item)
-        return "Oops, you dropped something!", False
+    # def _drop_item(self, item):
+    #     """drops specific item from player to room"""
+    #     player = self.hero.items
+    #     room_items = self.hero.location.items
+    #     room_items[item] = player.pop(item)
+    #     return "Oops, you dropped something!", False
 
-    def _use_item(self, item="Nothing"):
-        if item in self.hero.items.keys():
-            target = input("What do you want to use it on? ")
-            self.hero.items[item].use(target, item)
-            return "You used the thing! It's super effective!", False
-        return "You don't have that item...", False
+    # def _use_item(self, item="Nothing"):
+    #     if item in self.hero.items.keys():
+    #         target = input("What do you want to use it on? ")
+    #         self.hero.items[item].use(target, item)
+    #         return "You used the thing! It's super effective!", False
+    #     return "You don't have that item...", False
 
-    def _start_over(self, load_or_save):
+    def _start_over(self):
         if self._confirm():
-            out = "new game"
+            out = ""
         else:
             out = "Guess you changed your mind!"
         return out, False
